Refactor RFID Reader blocks to new module.
